day-1

+++ migrations/20230921045922_init.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS coffees (
     "name" VARCHAR NOT NULL,
     "roast" VARCHAR NOT NULL,
     "region" VARCHAR NOT NULL,
-    "price" DOUBLE NOT NULL,
+    "price" FLOAT NOT NULL,
     "grind_uint" INT NOT NULL,
     "created_at" TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     "updated_at" TIMESTAMP WITH TIME ZONE DEFAULT NOW()
